about halfway through merging entry points to RBM into a single file
