+++ systems/plants/CMakeLists.txt
@@ -43,33 +43,21 @@ if (eigen3_FOUND)
     add_subdirectory(viewer)
   endif(Boost_FOUND)
 
-  macro(add_rbm_mex)
-    # todo: use this again once I can assume everyone has CMAKE version >= 2.8.8
-#    add_mex(${ARGV} ${ARGV}.cpp $<TARGET_OBJECTS:drakeUtil>)
-#    target_link_libraries(${ARGV} drakeRBM)
-    add_mex(${ARGV} ${ARGV}.cpp)
-    target_link_libraries(${ARGV} drakeRBM drakeUtil)
-  endmacro()
+  add_mex(rigidBodyManipulatormex rigidBodyManipulatormex.cpp)
+  target_link_libraries(rigidBodyManipulatormex drakeRBM drakeUtil)
 
-  add_rbm_mex(constructModelmex)
-  add_rbm_mex(HandCmex)
-  add_rbm_mex(doKinematicsmex)
-  add_rbm_mex(forwardKinmex)
-  add_rbm_mex(bodyKinmex)
-  add_rbm_mex(collisionDetectmex)
-  add_rbm_mex(smoothDistancePenaltymex)
-  add_rbm_mex(collisionRaycastmex)
-  add_rbm_mex(allCollisionsmex)
-  add_rbm_mex(getCMMmex)
-  add_rbm_mex(findKinematicPathmex)
-  add_rbm_mex(geometricJacobianmex)
-  add_rbm_mex(doKinematicsNewmex)
-  add_rbm_mex(forwardKinVmex)
-  add_rbm_mex(forwardKinPositionGradientmex)
-  add_rbm_mex(centerOfMassmex)
-  add_rbm_mex(centroidalMomentumMatrixmex)
-  add_rbm_mex(massMatrixmex)
-  add_rbm_mex(inverseDynamicsmex)
+#  add_rbm_mex(collisionRaycastmex)
+#  add_rbm_mex(allCollisionsmex)
+#  add_rbm_mex(getCMMmex)
+#  add_rbm_mex(findKinematicPathmex)
+#  add_rbm_mex(geometricJacobianmex)
+#  add_rbm_mex(doKinematicsNewmex)
+#  add_rbm_mex(forwardKinVmex)
+#  add_rbm_mex(forwardKinPositionGradientmex)
+#  add_rbm_mex(centerOfMassmex)
+#  add_rbm_mex(centroidalMomentumMatrixmex)
+#  add_rbm_mex(massMatrixmex)
+#  add_rbm_mex(inverseDynamicsmex)
 
   add_mex(surfaceTangentsmex surfaceTangentsmex.cpp)
